+++ packages/storefront/tests/vrt/specs/markdown.vrt.ts
@@ -1,6 +1,12 @@
 import { expect, test } from '@playwright/test';
-import { schemes, viewportWidthM, viewportWidths } from '@porsche-design-system/shared/testing/playwright.vrt';
-import { closeSidebars, resetAnimations } from '../helpers/helpers';
+import {
+  schemes,
+  viewportWidthM,
+  viewportWidthXL,
+  viewportWidths,
+} from '@porsche-design-system/shared/testing/playwright.vrt';
+import { viewportWidthL } from 'shared/src/testing/playwright.vrt';
+import { closeSidebars, resetAnimations, waitForImagesToBeLoaded } from '../helpers/helpers';
 
 test.describe('markdown', async () => {
   schemes.forEach((scheme) => {
@@ -10,16 +16,15 @@ test.describe('markdown', async () => {
       await page.emulateMedia({
         colorScheme: scheme,
       });
-      await page.goto('/markdown');
-      await resetAnimations(page);
-      await page.evaluate(() =>
-        (window as unknown as Window & { componentsReady: () => Promise<number> }).componentsReady()
-      );
-      await page.focus('a[href="https://designsystem.porsche.com/"]:not([title])');
+      await page.goto('/-/mdx');
       await page.setViewportSize({
         width: viewportWidthM,
         height: await page.evaluate(() => document.body.clientHeight),
       });
+      await resetAnimations(page);
+      await page.evaluate(() =>
+        (window as unknown as Window & { componentsReady: () => Promise<number> }).componentsReady()
+      );
       const screenshot = await page.screenshot({ fullPage: true });
       expect(screenshot).toMatchSnapshot(`markdown-${viewportWidthM}-scheme-${scheme}.png`);
     });
@@ -29,17 +34,15 @@ test.describe('markdown', async () => {
     .filter((x) => x !== viewportWidthM)
     .forEach((viewportWidth) => {
       test(`should have no visual regression for viewport ${viewportWidth}`, async ({ page }) => {
-        await page.goto('/markdown');
-        await resetAnimations(page);
-        await page.evaluate(() =>
-          (window as unknown as Window & { componentsReady: () => Promise<number> }).componentsReady()
-        );
-        await page.focus('a[href="https://designsystem.porsche.com/"]:not([title])');
+        await page.goto('/-/mdx');
         await page.setViewportSize({
           width: viewportWidth,
           height: await page.evaluate(() => document.body.clientHeight),
         });
-
+        await resetAnimations(page);
+        await page.evaluate(() =>
+          (window as unknown as Window & { componentsReady: () => Promise<number> }).componentsReady()
+        );
         await closeSidebars(page);
 
         const screenshot = await page.screenshot({ fullPage: true });