+++ packages/components/CHANGELOG.md
@@ -18,6 +18,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 - `Select Wrapper`: update of scroll position if dropdown is navigated with keyboard
   ([#3858](https://github.com/porsche-design-system/porsche-design-system/pull/3858))
+
+## [3.28.0-rc.3] - 2025-05-19
+
+### Added
+
+- `Styles`: theme for `tailwindcss` including utilities available under
+  `import { … } from '@porsche-design-system/components-{js|angular|react|vue}/tailwindcss';`
+  ([#3849](https://github.com/porsche-design-system/porsche-design-system/pull/3849))
+
+### Changed
+
+- Partials: `getInitialStyles` removed default styles for `h1`, `h2`, `h3`, `h4`, `h5`, `h6`, `p`, `b` and `strong`
+  ([#3849](https://github.com/porsche-design-system/porsche-design-system/pull/3849))
+
+#### Fixed
+
 - `Multi Select`: enable full a11y compliance
   ([#3819](https://github.com/porsche-design-system/porsche-design-system/pull/3819))
 - Accessibility: fix ARIA live announcements of form components status messages
@@ -30,7 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Canvas`: stealing focus from `Text Field Wrapper`
   ([#3843](https://github.com/porsche-design-system/porsche-design-system/pull/3843))
 
-### [3.28.0-rc.1] - 2025-04-14
+## [3.28.0-rc.1] - 2025-04-14
 
 ### Fixed